Stop using frameworks
Fixes #12
Fixes #22

For some reason, when using a Swift-based framework, it tries to load Swift runtime libraries from the local Xcode install, and if this Xcode install is outdated, the app crashes with error about missing symbols. I have not found a way around this, so I'm just vendoring the few dependencies we have. Swift 5 will be ABI stable and bundle the Swift runtime libraries in the OS, so we can switch back to frameworks then.