Fix off-by-one error and ensure Gifski completion handler is only called once (#56)
I introduced an off-by-one error in 8a764ad695b097cf99356735a9653dd11f0ab1ef. It was only noticeable when trying to convert the video in https://github.com/sindresorhus/gifski-app/issues/46, which is why I didn't discover it when I did the change.

We also had the problem that the `completionHandler` could be called multiple times because we have it in the `generateCGImagesAsynchronously` closure which is called for each finished image or cancellation.