Add fromX to CALayer animateScaleMove
fromX and fromY are now optional and default to the center of the
layer. You can either jsut use one or even both values. Can be tested by
changing the only call to this function in this project
(ConversionCompletedViewController.swift#L60) to something
like that:

draggableFile.layer?.animateScaleMove(
	fromScale: 3,
	fromX: view.frame.width + draggableFile.frame.size.width,
	fromY: view.frame.height + draggableFile.frame.size.height
)

which actually looks nice.