fix: looping Controls UI flow
Prior to this change, UI for looping was not controlled very well

This change only allows positive numbers for the loop Counter field, checking the "Forever" box disables the loop counter text field and resets the value to 0, unchecking the forever box enables the loop counter text field. UI is also updated for the looping controls.