feat: stepper control for loop count
Prior to this change, the only way to change the number of loops was to type in a value.

This change adds in a stepper control to add a different way to change the number of loops of an animation. The stepper control is disabled and enabled at the same time as the loop counter textfield.