fix: loop controls
Prior to this change, the upper limit of looping was 9999 and the loop counter text box allowed negative numbers if arrow keys were used to change the value.

This change changes the looping range to 0-100 and clamps the values to that range. This change also has UI updates as suggested by code review.