fix: ui updates and ease of change modifications
Prior to this change, the loop/repeat count had to be changed in two places.

This change makes changing the looping count a one stop change. Also tweaks the UI to right align the Loop Forever checkbox to the sliders right side.

fix: UI element locations

Prior to this change, the use of "Add missing constraints" shifted the locations of multiple UI elements.

This change reverts those shifts in UI element positions.