fix: code review updates
Prior to this change, the location of the stepper was not aligned to the top of the loop count control.  The variable name for the repeatCountRange did not match the rest of the code.

This change fixes the UI location of the stepper and renames a variable.
Also removes a Rust warning about settings as mutable when it doesn't need to be.